refactor: refactor the code and make it cleaner
